finished

+++ Backend/src/controllers/notes.ts
@@ -2,10 +2,17 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const getAuthenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(getAuthenticatedUserId);
+
+    const notes = await NoteModel.find({
+      userId: getAuthenticatedUserId,
+    }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -14,8 +21,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
-
+  const getAuthenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(getAuthenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note ID");
     }
@@ -24,6 +33,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(getAuthenticatedUserId)) {
+      throw createHttpError(403, "You are not authorized to access this note");
     }
 
     res.status(200).json(note);
@@ -45,13 +58,17 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const getAuthenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(getAuthenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Title is required");
     }
 
     const newNote = await NoteModel.create({
+      userId: getAuthenticatedUserId,
       title: title,
       text: text,
     });
@@ -80,8 +97,11 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const getAuthenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(getAuthenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note ID");
     }
@@ -94,6 +114,9 @@ export const updateNote: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(getAuthenticatedUserId)) {
+      throw createHttpError(403, "You are not authorized to access this note");
     }
 
     note.title = newTitle;
@@ -109,8 +132,10 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const getAuthenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(getAuthenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note ID");
     }
@@ -119,6 +144,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(getAuthenticatedUserId)) {
+      throw createHttpError(403, "You are not authorized to access this note");
     }
 
     await note.deleteOne();
